move replace old pages to new components
