+++ pages/_old/index.tsx
@@ -1,0 +1,11 @@
+import { useA2HS } from '../../hooks/useA2HS'
+
+const Page = () => {
+  return (
+    <>
+      <div className="App">machiwabi</div>
+    </>
+  )
+}
+
+export default Page
