--- pages/index.tsx
@@ -1,11 +1,0 @@
-import { useA2HS } from '../hooks/useA2HS'
-
-const Page = () => {
-  return (
-    <>
-      <div className="App">machiwabi</div>
-    </>
-  )
-}
-
-export default Page
